Beginning infrastructure changes to migrate to mocha
