--- packages/v-connection-string/.coveralls.yml
@@ -1,2 +1,0 @@
-service_name: travis-pro
-repo_token: 5F6dODinz9L9uFR6HatKmtsYDoV1A5S2N
