--- packages/v-connection-string/.travis.yml
@@ -1,6 +1,0 @@
-language: node_js
-node_js:
-  - '0.10'
-  - '6.9'
-  - '8'
-after_success: 'npm run coveralls'
